+++ util/random.go
@@ -14,8 +14,8 @@ func init() {
 	rand.Seed(time.Now().UnixNano())
 }
 
-func RandomInt(min, max int64) int64 {
-	return min + rand.Int63n(max-min+1)
+func RandomInt(min, max int64) float64 {
+	return float64(min + rand.Int63n(max-min+1))
 }
 
 func RandomString(n int) string {
@@ -34,7 +34,7 @@ func RandomOwner() string {
 	return RandomString(6)
 }
 
-func RandomMoney() int64 {
+func RandomMoney() float64 {
 	return RandomInt(0, 1000)
 }
